Add a CSV download option for the site inventory.
While this allows a CSV to be generated on the source site, the long-
term goal is to be able to use migrate.module to pull in data from
multiple source sites. As a result, we need to make the CSV files
accessible without being authenticated to the site. We do this by
removing all Views access checks.

However, we still want to secure access to the site inventory, as it
may contain sensitive information. So, we add a custom access callback
that checks for a key passed as a query parameter.